Bootstrappified the main status-posting page
The posting functionality still works, but the error message is disabled temporarily. Need to figure out how to get the filter checkboxes all on the same line and investigate if there's a special bootstrap checkbox. Also need to figure out how to incorporate the #error-alert/#error-message without the page breaking. Need to rearrange the layout of the page so it looks better too.